new set of casts missed before

+++ lib/ts/TextBuffer.h
@@ -85,7 +85,7 @@ public:
   size_t
   spaceUsed() const
   {
-    return (size_t)(nextAdd - bufferStart);
+    return nextAdd - bufferStart;
   };
 
   bool
